[0.0.1]🔧Fix: 修复了部分安卓依赖问题

+++ swarm_cloud/android/src/main/java/com/swarmcloud/sdk/CdnbyeMethodHandler.java
@@ -96,7 +96,7 @@ public class CdnbyeMethodHandler implements MethodChannel.MethodCallHandler {
             if(configMap.get("httpHeaders") != null)
                 builder = builder.httpHeadersForHls((Map<String, String>) configMap.get("httpHeaders"));
 
-            builder = builder.httpLoadTime((long) configMap.get("httpLoadTime"));
+            builder = builder.httpLoadTime((int) configMap.get("httpLoadTime"));
             builder = builder.logPersistent((boolean) configMap.get("logPersistent"));
             builder = builder.sharePlaylist((boolean) configMap.get("sharePlaylist"));
             builder = builder.waitForPeer((boolean) configMap.get("waitForPeer"));
